Extract executeFfmpegInChildProcess to its own file - Enable mocking it easily - Separate executeFfmpegInChildProcess dependencies from generateThumbnail dependencies
